move metrics test to a new file
